refactor(web): lift formatFileSize to the shared utils it is shared from
Moving the bundle confirm page into `components/` left a top-level shared
component importing `@/domains/workspace/utils/format-file-size`, which
inverts the dependency direction `docs/CONVENTIONS.md` states as one-way:
`shared -> domains -> page domains -> routes`. Shared code does not reach
into a domain.

`formatFileSize` is a pure function with no dependencies, so it matches
the `utils/` half of that doc's lib-vs-utils table exactly, and its three
callers now span a domain and a shared component. It moves to
`src/utils/format-file-size.ts` with its test beside it, which is where
the decision rule for shared code puts it.

`src/domains/workspace/utils/` keeps its four other files, so nothing is
left as an empty or single-file directory, and the generated
`.cross-domain-allowlist.json` is unchanged.
